+++ src/config_file.cpp
@@ -236,16 +236,20 @@ const std::string &config_file::get(const std::string &variable, bool silent) co
     return empty;
 }
 
+std::string config_file::interpret_config_path(const std::string& path) const {
+    return filesystem::absolute(path, filename.parent_path()).u8string();
+}
+
 std::string config_file::get_path(const std::string& variable, bool silent) const {
     auto str = get(variable, silent);
-    return filesystem::absolute(str, filename.parent_path()).u8string();
+    return interpret_config_path(str);
 }
 
 std::vector<std::string> config_file::get_paths(const std::string& variable, bool silent) const {
     auto strs = get_vector(variable, silent);
 
     for (size_t i = 0; i < strs.size(); ++i) {
-        strs[i] = filesystem::absolute(strs[i], filename.parent_path()).u8string();
+        strs[i] = interpret_config_path(strs[i]);
     }
     return strs;
 }
@@ -280,7 +284,7 @@ std::string config_file::require_path(const std::string &variable) const {
         return "";
     }
 
-    return filesystem::absolute(str, filename.parent_path()).u8string();
+    return interpret_config_path(str);
 }
 
 //static size_t send_write_func(void *ptr, size_t size, size_t nmemb, void *userdata) {