+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     is now limited to only ES256 (which is what Kontosynk uses). (#103)
   - Support for filtering out orphans after the load phase (#108)
   - Load limiting can now be done with regular expressions (#110)
+  - Load limiters can now be combined (#112)
 
 #### Bugfixes
   - Load limiting now also applies to auxilliary attributes (#99)